requirements.txt
Flask>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
redis>=4.0
SQLAlchemy>=1.4
mysql-connector-python>=8.0
requests>=2.0
gunicorn>=20.1
